Checkpoint: Implement new bucket system
This commit does not include unit tests and isn't a use candidate.

The bucket system has been overhauled. It now uses a proxy pattern,
with the exported Bucket class supporting a variety of backends.
The stock implementations of these are an adaption of the previously
existing FileBucket, as well as the new DatabaseBackend.

The updated Bucket class should have identical behaviour to its
predecessor with the expection of providing more options, emitting
deprecation warnings, and being more explicity about key errors.