Initial CLI refactor
Added an architectural layer to the CLI and refactored it to be
package-level. CLI commands now support more complex behaviour,
including wizarding and option type assertion, with simple usage
code.

Note this implementation doesn't use docopt as docopt doesn't allow
options wildcards and I wasn't able to construct a usage descriptor
it would reliably parse with all options included.

Code in this implementation that we want to keep needs to be commented.