Enable reading of JSON request bodies.
This change replaces a previous error case, where a supplied JSON
body would be passed to the cgi.FieldStorage for parsing, which
would die. The zoom.request.Request class has also been modified
to be Content-Type and Accept aware. Request().data will now return
an empty dict in the case that the client supplied something other
than a form in the request body.

JSON is available in Request().json_body, which is lazy loaded and
multi-access safe (this property already existed but was
unreachable).

This branch also fixes an issue found in development configurations
where reading the WSGI request body input stream provided by Pythons
simple server will block indefinitely if the request body is not
terminated by newlines.